Switched relation between Category and MenuItem. It allows for displaying whole menu in elegant manner on client side in React

+++ src/main/java/com/hackybear/hungry_scan_core/service/CategoryServiceImp.java
@@ -50,6 +50,10 @@ public class CategoryServiceImp implements CategoryService {
     @Override
     public void delete(Integer id) throws LocalizedException {
         Category existingCategory = findById(id);
-        categoryRepository.delete(existingCategory);
+        if (!existingCategory.getMenuItems().isEmpty()) {
+            exceptionHelper.throwLocalizedMessage("error.categoryService.categoryNotEmpty");
+        } else {
+            categoryRepository.delete(existingCategory);
+        }
     }
 }
